chore: Rename Tooltip to LegacyTooltip (#1518)
