+++ src/components/LegacyTooltip/index.ts
@@ -1,4 +1,4 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './BrightHeader';
-export * from './Tooltip';
+export * from './LegacyTooltip';
